+++ modules/canvaskit/tests/paragraph.spec.js
@@ -73,7 +73,7 @@ describe('Paragraph Behavior', function() {
             decorationThickness: 1.5, // multiplier based on font size
             fontSize: 24,
         });
-        builder.pushStyle(blueText)
+        builder.pushStyle(blueText);
         builder.addText(`Gosh I hope this wraps at some point, it is such a long line.`)
         builder.pop();
         builder.addText(` I'm done with the blue now. `)
@@ -108,7 +108,44 @@ describe('Paragraph Behavior', function() {
         builder.delete();
     });
 
-        // loosely based on SkParagraph_GetRectsForRangeParagraph test in c++ code.
+    gm('paragraph_foreground_and_background_color', (canvas) => {
+        const paint = new CanvasKit.SkPaint();
+
+        paint.setColor(CanvasKit.RED);
+        paint.setStyle(CanvasKit.PaintStyle.Stroke);
+
+        const fontMgr = CanvasKit.SkFontMgr.FromData(notoSerifFontBuffer);
+        expect(fontMgr.countFamilies()).toEqual(1);
+        expect(fontMgr.getFamilyName(0)).toEqual('Noto Serif');
+
+        const wrapTo = 200;
+
+        const paraStyle = new CanvasKit.ParagraphStyle({
+            textStyle: {
+                foregroundColor: CanvasKit.Color4f(1.0, 0, 0, 0.8),
+                backgroundColor: CanvasKit.Color4f(0, 0, 1.0, 0.8),
+                // color should default to black
+                fontFamilies: ['Noto Serif'],
+                fontSize: 20,
+            },
+
+            textAlign: CanvasKit.TextAlign.Center,
+        });
+        const builder = CanvasKit.ParagraphBuilder.Make(paraStyle, fontMgr);
+        builder.addText(
+            'This text has a red foregroundColor and a blue backgroundColor.');
+        const paragraph = builder.build();
+        paragraph.layout(300);
+
+        canvas.drawParagraph(paragraph, 10, 10);
+
+        paint.delete();
+        fontMgr.delete();
+        paragraph.delete();
+        builder.delete();
+    });
+
+    // loosely based on SkParagraph_GetRectsForRangeParagraph test in c++ code.
     gm('paragraph_rects', (canvas) => {
         const fontMgr = CanvasKit.SkFontMgr.FromData(notoSerifFontBuffer);
 
@@ -116,9 +153,12 @@ describe('Paragraph Behavior', function() {
         const hStyle = CanvasKit.RectHeightStyle.Max;
         const wStyle = CanvasKit.RectWidthStyle.Tight;
 
+        const mallocedColor = CanvasKit.Malloc(Float32Array, 4);
+        mallocedColor.toTypedArray().set([0.9, 0.1, 0.1, 1.0]);
+
         const paraStyle = new CanvasKit.ParagraphStyle({
             textStyle: {
-                color: CanvasKit.BLACK,
+                color: mallocedColor,
                 fontFamilies: ['Noto Serif'],
                 fontSize: 50,
             },
@@ -128,6 +168,7 @@ describe('Paragraph Behavior', function() {
         const builder = CanvasKit.ParagraphBuilder.Make(paraStyle, fontMgr);
         builder.addText('12345,  \"67890\" 12345 67890 12345 67890 12345 67890 12345 67890 12345 67890 12345');
         const paragraph = builder.build();
+        CanvasKit.Free(mallocedColor);
 
         paragraph.layout(wrapTo);
 
@@ -385,7 +426,7 @@ describe('Paragraph Behavior', function() {
                 slant: CanvasKit.FontSlant.Italic,
             }
         });
-        builder.pushStyle(boldItalic)
+        builder.pushStyle(boldItalic);
         builder.addText(`Bold, Italic\n`); // doesn't show up, but we don't crash
         builder.pop();
         builder.addText(`back to normal`);
@@ -407,7 +448,7 @@ describe('Paragraph Behavior', function() {
 
     it('should not crash if we omit font family on paragraph style', () => {
         const surface = CanvasKit.MakeCanvasSurface('test');
-        expect(surface).toBeTruthy('Could not make surface')
+        expect(surface).toBeTruthy('Could not make surface');
 
         const canvas = surface.getCanvas();
         const paint = new CanvasKit.SkPaint();
@@ -436,7 +477,7 @@ describe('Paragraph Behavior', function() {
                 slant: CanvasKit.FontSlant.Italic,
             }
         });
-        builder.pushStyle(boldItalic)
+        builder.pushStyle(boldItalic);
         builder.addText(`Bold, Italic\n`);
         builder.pop();
         builder.addText(`back to normal`);